requirements.txt
# ============================================================
#  CogniView.AI — Python Dependencies
#  Install with:  pip install -r requirements.txt
# ============================================================

# ── Core Web Framework ──────────────────────────────────────
fastapi>=0.110.0                 # REST API framework
uvicorn>=0.27.0                  # ASGI server to run FastAPI
python-multipart>=0.0.9          # Required by FastAPI for file uploads (UploadFile)
pydantic>=2.0.0                  # Data validation & schemas

# ── HTTP Clients ────────────────────────────────────────────
httpx>=0.27.0                    # Async HTTP client (LLM calls, API requests)
requests>=2.31.0                 # Sync HTTP client (database searches, PDF downloads)

# ── PDF Processing ──────────────────────────────────────────
PyMuPDF>=1.24.0                  # PDF text extraction & parsing (imported as 'fitz')

# ── Environment & Config ────────────────────────────────────
python-dotenv>=1.0.0             # Load .env files for API keys & configuration

# ── NLP & Keyword Extraction ────────────────────────────────
spacy>=3.7.0                     # NLP pipeline for keyword extraction
keybert>=0.8.0                   # Keyword extraction using BERT embeddings
sentence-transformers>=2.2.0     # Sentence embeddings (used by KeyBERT & semantic search)

# ── SPECTER 2.0 Semantic Ranking ────────────────────────────
adapters>=1.0.0                  # Adapter-Transformers for SPECTER 2.0 model
transformers>=4.38.0             # Hugging Face Transformers (AutoTokenizer)
torch>=2.1.0                     # PyTorch backend for transformer models
numpy>=1.24.0                    # Numerical computations (cosine similarity, normalization)

# ── XML Parsing ─────────────────────────────────────────────
lxml>=4.9.0                      # Fast XML/HTML parser (used in database searcher)

# ============================================================
#  After installing Python dependencies, also install:
#
#  1. spaCy language model:
#     python -m spacy download en_core_web_sm
#
#  2. Node.js dependencies (for the React frontend):
#     npm install
#
#  3. Create a .env file in src/ with your API keys
#     (see README.md for required environment variables)
# ============================================================